#321 coalesce model elements into sis.elements
