--- student-info-system/professor/forms.py
@@ -1,10 +1,0 @@
-from django.forms import ModelForm
-
-from sis.models import ReferenceItem
-
-
-class ReferenceItemForm(ModelForm):
-
-    class Meta:
-        model = ReferenceItem
-        fields = ['title', 'description', 'link', 'edition', 'isbn', 'type']
