+++ student-info-system/schooladmin/views/__init__.py
@@ -1,12 +1,12 @@
 from datetime import date, datetime
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.db.models import Q, Subquery
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
-from django.utils.html import format_html
 from django.urls import reverse
-from django.contrib.auth.decorators import login_required
-from django.db.models import Q, Subquery
+from django.utils.html import format_html
 
 from sis.authentication_helpers import role_login_required
 
@@ -15,33 +15,20 @@ from sis.models import (Course, CoursePrerequisite, Major, Professor, Section, S
 
 from sis.utils import filtered_table, filtered_table2, DUMMY_ID, next_prev
 
-from sis.filters.course import CourseFilter
-from sis.filters.section import SectionFilter
-from sis.filters.sectionreferenceitem import SectionItemFilter
-from sis.filters.sectionstudent import SectionStudentFilter
-from sis.filters.semester import SemesterFilter
-from sis.filters.user import StudentFilter, UserFilter, ProfessorFilter
-from sis.tables.referenceitems import ProfReferenceItemsTable
-
-from schooladmin.forms import (
-    CourseCreationForm,
-    CourseEditForm,
-    SemesterCreationForm,
-    SemesterEditForm,
-)
-
-from sis.forms.major import MajorCreationForm, MajorEditForm
-from sis.forms.section import SectionCreationForm, SectionEditForm
-from sis.forms.referenceitem import ReferenceItemCreationForm
-from sis.filters.referenceitem import ItemFilter
-
-from sis.tables.courses import CoursesTable, CoursesForMajorTable, MajorCoursesMetTable
-from sis.tables.sectionreferenceitems import ReferenceItemsForSectionTable
-from sis.tables.sections import SectionForClassTable, SectionsTable
-from sis.tables.sectionstudents import (StudentHistoryTable, StudentInSectionTable)
-from sis.tables.semesters import SemestersSummaryTable, SemestersTable
-from sis.tables.users import (UsersTable, FullUsersTable, StudentsTable, StudentInMajorTable,
-                              ProfessorsTable)
+from sis.elements.course import (CoursesTable, CoursesForMajorTable, MajorCoursesMetTable,
+                                 CourseCreationForm, CourseEditForm, CourseFilter)
+from sis.elements.major import MajorCreationForm, MajorEditForm
+from sis.elements.referenceitem import (ItemFilter, ProfReferenceItemsTable,
+                                        ReferenceItemCreationForm)
+from sis.elements.section import (SectionForClassTable, SectionsTable, SectionFilter,
+                                  SectionCreationForm, SectionEditForm)
+from sis.elements.sectionreferenceitem import SectionItemFilter, ReferenceItemsForSectionTable
+from sis.elements.sectionstudent import (StudentHistoryTable, StudentInSectionTable,
+                                         SectionStudentFilter)
+from sis.elements.semester import (SemestersSummaryTable, SemestersTable, SemesterCreationForm,
+                                   SemesterEditForm, SemesterFilter)
+from sis.elements.user import (UsersTable, FullUsersTable, StudentsTable, StudentInMajorTable,
+                               ProfessorsTable, StudentFilter, UserFilter, ProfessorFilter)
 
 from easy_pdf import rendering
