+++ student-info-system/schooladmin/views/users.py
@@ -2,39 +2,34 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db import transaction
+from django.db.models import Exists, OuterRef
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.urls import reverse
-from django.db.models import Exists, OuterRef
 
 from sis.authentication_helpers import role_login_required
 
 from sis.models import (Course, CoursePrerequisite, Major, Professor, Section, Semester, Student,
                         SectionStudent, Profile, Message)
 
-from sis.forms.profile import DemographicForm, ProfileCreationForm, ProfileEditForm
-from sis.forms.student import StudentEditForm, StudentCreationForm
-from sis.forms.user import UserCreationForm, UserEditForm
-from sis.forms.professor import ProfessorCreationForm, ProfessorEditForm
-
 from sis.utils import filtered_table2, DUMMY_ID, next_prev
-from sis.filters.course import CourseFilter
-from sis.filters.message import (FullSentMessageFilter, FullReceivedMessageFilter,
-                                 SentMessageFilter, ReceivedMessageFilter)
-from sis.filters.referenceitem import ItemFilter
-from sis.filters.section import SectionFilter
-from sis.filters.sectionstudent import SectionStudentFilter
-from sis.filters.semester import SemesterFilter
-from sis.filters.user import StudentFilter, UserFilter, ProfessorFilter
-from sis.tables.messages import MessageSentTable, MessageReceivedTable
-from sis.tables.referenceitems import ProfReferenceItemsTable
 
-from sis.tables.courses import CoursesTable, CoursesForMajorTable, MajorCoursesMetTable
-from sis.tables.sections import SectionForClassTable, SectionsTable
-from sis.tables.sectionstudents import (StudentHistoryTable, StudentInSectionTable)
-from sis.tables.semesters import SemestersSummaryTable, SemestersTable
-from sis.tables.users import (UsersTable, FullUsersTable, StudentsTable, StudentInMajorTable,
-                              ProfessorsTable)
+from sis.elements.course import (CourseFilter, CoursesTable, CoursesForMajorTable,
+                                 MajorCoursesMetTable)
+from sis.elements.message import (FullSentMessageFilter, FullReceivedMessageFilter,
+                                  SentMessageFilter, ReceivedMessageFilter, MessageSentTable,
+                                  MessageReceivedTable)
+from sis.elements.professor import ProfessorCreationForm, ProfessorEditForm
+from sis.elements.profile import DemographicForm, ProfileCreationForm, ProfileEditForm
+from sis.elements.referenceitem import ItemFilter, ProfReferenceItemsTable
+from sis.elements.section import SectionFilter, SectionForClassTable, SectionsTable
+from sis.elements.sectionstudent import (StudentHistoryTable, StudentInSectionTable,
+                                         SectionStudentFilter)
+from sis.elements.semester import SemesterFilter, SemestersSummaryTable, SemestersTable
+from sis.elements.student import StudentEditForm, StudentCreationForm
+from sis.elements.user import (StudentFilter, UserFilter, ProfessorFilter, UserCreationForm,
+                               UserEditForm, UsersTable, FullUsersTable, StudentsTable,
+                               StudentInMajorTable, ProfessorsTable)
 
 
 @role_login_required(Profile.ACCESS_ADMIN)
