--- student-info-system/sis/forms/user.py
@@ -1,25 +1,0 @@
-from django import forms
-from django.contrib.auth.forms import UserCreationForm as DjangoUserCreationForm
-from django.contrib.auth.models import User
-
-from sis.forms.utils import *
-
-
-class UserCreationForm(DjangoUserCreationForm):
-    first_name = forms.CharField(max_length=30, required=True, help_text='Required.')
-    last_name = forms.CharField(max_length=30, required=True, help_text='Required.')
-    email = forms.EmailField(max_length=254, help_text='Required. Enter a valid email address.')
-
-    class Meta:
-        model = User
-        fields = ('username', 'first_name', 'last_name', 'email', 'password1', 'password2')
-
-
-class UserEditForm(forms.ModelForm):
-    first_name = forms.CharField(max_length=30, required=True, help_text='Required.')
-    last_name = forms.CharField(max_length=30, required=True, help_text='Required.')
-    email = forms.EmailField(max_length=254, help_text='Required. Enter a valid email address.')
-
-    class Meta:
-        model = User
-        fields = ('first_name', 'last_name', 'email')
