+++ student-info-system/sis/views/__init__.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
 from django.shortcuts import redirect, render
+
 from sis.models import Profile
