+++ student-info-system/sis/views/sectionreferenceitem.py
@@ -1,10 +1,7 @@
-from django.contrib import messages
-from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
-from django.shortcuts import redirect, render
-from django.urls import reverse
+from django.shortcuts import render
 
-from sis.models import (SectionReferenceItem)
+from sis.models import SectionReferenceItem
 
 
 def sectionreferenceitem(request, id):
